+++ src/snowflake/cli/_app/printing.py
@@ -17,7 +17,8 @@ from __future__ import annotations
 import csv
 import json
 import sys
-from datetime import date, datetime
+from datetime import date, datetime, time
+from decimal import Decimal
 from json import JSONEncoder
 from pathlib import Path
 from textwrap import indent
@@ -58,9 +59,9 @@ class CustomJSONEncoder(JSONEncoder):
             return o.result
         if isinstance(o, (CollectionResult, MultipleResults)):
             return list(o.result)
-        if isinstance(o, (date, datetime)):
+        if isinstance(o, (date, datetime, time)):
             return o.isoformat()
-        if isinstance(o, Path):
+        if isinstance(o, (Path, Decimal)):
             return str(o)
         return super().default(o)